+++ src/test/kotlin/lotto/PrizeTest.kt
@@ -2,7 +2,7 @@ package lotto
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
-import lotto.common.Money
+import lotto.common.value.Money
 import lotto.domain.enums.Prize
 
 class PrizeTest : FreeSpec({
